+++ lib/src/widgets/widgets.dart
@@ -1,5 +1,6 @@
 export '../snackbars/snackbars.dart';
 export 'column_with_space.dart';
+export 'hide_keyboard.dart';
 export 'row_with_space.dart';
 export 'scaffold_wrapper.dart';
 export 'social_button.dart';